requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
h5py==3.7.0
huggingface-hub==0.9.1
ipython==8.6.0
keras==2.9.0
Keras-Preprocessing==1.1.2
matplotlib==3.6.0
numpy==1.23.2
opencv-python==4.6.0.66
pandas==1.5.1
Pillow==9.2.0
pydantic==1.9.2
pydicom==2.3.0
requests==2.28.1
seaborn==0.12.1
tensorflow==2.9.2
torch==1.13.0
torchvision==0.14.0
transformers==4.22.0